Update build factories to build against recent SDK
With SFML/SFML#1214 & SFML/SFML#1215, cmake will by default use the oldest SDK that we support. This patch change the last build step to build against the newest SDK available on our Mac mini.

NOTE: when the Mac mini is upgraded to 10.12 (or later), this needs to be updated as well.